Rename to toEthereumSnowbridgeV2
