+++ web/packages/api/src/index.ts
@@ -254,4 +254,4 @@ export * as subsquid from "./subsquid"
 export * as forKusama from "./forKusama"
 export * as toEthereumFromEVMV2 from "./toEthereumFromEVM_v2"
 export * as parachains from "./parachains"
-export * as toEthereumV3 from "./toEthereum_v3"
+export * as toEthereumSnowbridgeV2 from "./toEthereumSnowbridgeV2"
