+++ web/packages/operations/package.json
@@ -37,13 +37,13 @@
     "transferXQCYFromEthereumToAH": "npx ts-node src/transfer_from_e2p.ts 1000 xrqcy 100000",
     "registerWeth": "npx ts-node src/register_erc20.ts 0xb8ea8cb425d85536b158d661da1ef0895bb92f1d",
     "transferWethToAH": "npx ts-node src/transfer_from_e2p.ts 1000 WETH 200000000000000",
-    "transferWndToEthereumV3": "npx ts-node src/transfer_from_p2e_v3.ts 1000 WND 2000000000",
-    "transferEtherFromAHV3": "npx ts-node src/transfer_from_p2e_v3.ts 1000 Eth 100000000000000",
-    "transferWEthFromAHV3": "npx ts-node src/transfer_from_p2e_v3.ts 1000 WETH 100000000000000",
-    "transferRocToEthereumV3": "npx ts-node src/transfer_from_p2e_v3.ts 1000 Roc 2000000000",
-    "transferPenpalTokenToEthereumV3": "npx ts-node src/transfer_from_p2e_v3.ts 2000 pal 2000000000",
-    "transferEtherFromPenpalV3": "npx ts-node src/transfer_from_p2e_v3.ts 2000 ETH 100000000000000",
-    "transferWEthFromPenpalV3": "npx ts-node src/transfer_from_p2e_v3.ts 2000 WETH 100000000000000"
+    "transferWndToEthereumV2": "npx ts-node src/transfer_from_p2e_v2.ts 1000 WND 2000000000",
+    "transferEtherFromAHV2": "npx ts-node src/transfer_from_p2e_v2.ts 1000 Eth 100000000000000",
+    "transferWEthFromAHV2": "npx ts-node src/transfer_from_p2e_v2.ts 1000 WETH 100000000000000",
+    "transferRocToEthereumV2": "npx ts-node src/transfer_from_p2e_v2.ts 1000 Roc 2000000000",
+    "transferPenpalTokenToEthereumV2": "npx ts-node src/transfer_from_p2e_v2.ts 2000 pal 2000000000",
+    "transferEtherFromPenpalV2": "npx ts-node src/transfer_from_p2e_v2.ts 2000 ETH 100000000000000",
+    "transferWEthFromPenpalV2": "npx ts-node src/transfer_from_p2e_v2.ts 2000 WETH 100000000000000"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.62.0",
