+++ web/packages/operations/src/transfer_from_p2e_v2.ts
@@ -1,5 +1,5 @@
 import "dotenv/config"
-import { transferToEthereum } from "./transfer_to_ethereum_v3"
+import { transferToEthereum } from "./transfer_to_ethereum_v2"
 
 const transfer = async (sourceParaId: number, symbol: string, amount: bigint) => {
     await transferToEthereum(sourceParaId, symbol, amount)
